restructure secrets/backend package
