+++ pkg/controller/externalsecret/externalsecret_functions_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/ContainerSolutions/externalsecret-operator/pkg/apis/externalsecretoperator/v1alpha1"
-	"github.com/ContainerSolutions/externalsecret-operator/secrets"
+	"github.com/ContainerSolutions/externalsecret-operator/secrets/backend"
 	"github.com/ContainerSolutions/externalsecret-operator/secrets/dummy"
 	. "github.com/smartystreets/goconvey/convey"
 )
@@ -13,9 +13,9 @@ func TestNewSecretForCR(t *testing.T) {
 	key := "key"
 	suffix := "-value"
 
-	secrets.Register("dummy", dummy.NewBackend)
-	secrets.Instantiate("dummy", "dummy")
-	secrets.Instances["dummy"].Init(map[string]string{"suffix": "-value"})
+	backend.Register("dummy", dummy.NewBackend)
+	backend.Instantiate("dummy", "dummy")
+	backend.Instances["dummy"].Init(map[string]string{"suffix": "-value"})
 
 	Convey("Given an ExternalSecret resource", t, func() {
 		externalSecret := v1alpha1.ExternalSecret{
