+++ secrets/asm/backend.go
@@ -4,7 +4,7 @@ package asm
 import (
 	"fmt"
 
-	"github.com/ContainerSolutions/externalsecret-operator/secrets"
+	"github.com/ContainerSolutions/externalsecret-operator/secrets/backend"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -20,11 +20,11 @@ type Backend struct {
 }
 
 func init() {
-	secrets.Register("asm", NewBackend)
+	backend.Register("asm", NewBackend)
 }
 
 // NewBackend returns an uninitialized Backend for AWS Secret Manager
-func NewBackend() secrets.Backend {
+func NewBackend() backend.Backend {
 	return &Backend{}
 }
