+++ secrets/backend/backend_test.go
@@ -1,4 +1,4 @@
-package secrets
+package backend
 
 import (
 	"encoding/json"
